requirements.txt
click==8.1.3
colorama==0.4.6
coverage==7.2.5
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
python-dotenv==1.0.0
tomli==2.0.1
